pyproject.toml
[tool.poetry]
name = "Contacts"
version = "3.1.0"
description = ""
authors = ["UstymenkoOB <50136631+UstymenkoOB@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.21"
psycopg2 = "^2.9.9"
alembic = "^1.12.0"
pydantic = "^2.4.2"
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
redis = "4.6.0"
fastapi-limiter = "0.1.5"
cloudinary = "^1.36.0"
pytest = "^7.4.3"



[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
